requirements.txt
Django==1.7.5
django-braces==1.4.0
django-crispy-forms==1.4.0
django-geojson==2.6.0
django-leaflet==0.15.2
django-materialize-css==0.0.1
django-materializecss-form==0.0.64
django-social-auth==0.7.28
django-uuidfield==0.5.0
httplib2==0.9
oauth2==1.5.211
oauthlib==0.7.2
Pillow==2.7.0
psycopg2==2.6
PyJWT==0.4.1
python-openid==2.2.5
python-social-auth==0.2.2
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
